+++ src/leiningen/new/db.clj
@@ -16,7 +16,7 @@
    ({:postgres ['org.postgresql/postgresql "42.1.4"]
      :mysql    ['mysql/mysql-connector-java "6.0.5"]
      :h2       ['com.h2database/h2 "1.4.196"]
-     :sqlite   ['org.xerial/sqlite-jdbc "3.20.0"]}
+     :sqlite   ['org.xerial/sqlite-jdbc "3.21.0.1"]}
      (select-db options))])
 
 (defn db-url [{:keys [sanitized] :as options} suffix]